% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRCIT.R
\name{.logrank_trafo2}
\alias{.logrank_trafo2}
\title{Logrank transformation function for LTRC data}
\usage{
.logrank_trafo2(x2)
}
\arguments{
\item{x2}{A vector \link[survival]{Surv} (Surv(time1, time2, event)) objects}
}
\value{
Logrank scores of LTRC objects
}
\description{
\code{.logrank_trafo} transforms Surv(time1, time2, event) objects into
logrank scores, which will be used later in the tree algorithm. It is
not designed to be used by users, not for internal used of \code{LTRCIT}
function.
}
