% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRCIT.R
\name{extree_data}
\alias{extree_data}
\title{Copy the partykit::extree_data function from partykit to avoid dependency issue}
\usage{
extree_data(
  formula,
  data,
  subset,
  na.action = stats::na.pass,
  weights,
  offset,
  cluster,
  strata,
  scores = NULL,
  yx = c("none", "matrix"),
  ytype = c("vector", "data.frame", "matrix"),
  nmax = c(yx = Inf, z = Inf),
  ...
)
}
\arguments{
\item{formula}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{data}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{subset}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{na.action}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{weights}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{offset}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{cluster}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{strata}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{scores}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{yx}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{ytype}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{nmax}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}

\item{...}{Same as the one in \link[partykit]{extree_data}, check \link[partykit]{extree_data} for usage}
}
\value{
check \link[partykit]{extree_data} for the return value
}
\description{
\code{extree_data} imports partykit::extree_data function
}
